requirements.txt
atomicwrites
attrs
certifi
charset-normalizer
colorama
cssselect
fake-useragent
idna
importlib-metadata
iniconfig
jsonpath
jsonschema
lxml
packaging
parse
pip
pluggy
py
PyHamcrest
pyparsing
pyquery
pyrsistent
pytest
PyYAML
requests
selenium
setuptools
six
toml
typing-extensions
urllib3
w3lib
xmljson
zipp
